requirements.txt
python-telegram-bot==20.7
openai==1.6.1
python-dotenv
aiohttp
uvloop